dcu: Fix build warnings when CONFIG_FSL_DCU_FB defined
Issue: ALB-3273
(cherry picked from commit cead89a623e348c78bb1e4a6607303083119057b)

+++ drivers/video/fsl_dcu_fb.c
@@ -369,7 +369,7 @@ void *video_hw_init(void)
 			      options + 8, fsl_dcu_mode_db) < 0)
 		return NULL;
 
-	ctfb.frameAdrs = (unsigned int)info.screen_base;
+	ctfb.frameAdrs = (unsigned int)(uintptr_t)info.screen_base;
 	ctfb.plnSizeX = ctfb.winSizeX;
 	ctfb.plnSizeY = ctfb.winSizeY;
